sumka

+++ src/ba/unsa/etf/rpr/Artikl.java
@@ -2,7 +2,7 @@ package ba.unsa.etf.rpr;
 
 public class Artikl {
     private String naziv;
-    private double cijena;
+    private int cijena;
     private String kod;
 
     public String getNaziv() {
@@ -13,11 +13,11 @@ public class Artikl {
         this.naziv = naziv;
     }
 
-    public double getCijena() {
+    public int getCijena() {
         return cijena;
     }
 
-    public void setCijena(double cijena) {
+    public void setCijena(int cijena) {
         this.cijena = cijena;
     }
 
@@ -29,7 +29,7 @@ public class Artikl {
         this.kod = kod;
     }
 
-    public Artikl(String naziv, double cijena, String kod){
+    public Artikl(String naziv, int cijena, String kod){
         this.naziv = naziv;
         this.cijena = cijena;
         this.kod = kod;
